Refactored group / resourcetype extension
